+++ src/components/UserCardProfile.jsx
@@ -4,7 +4,6 @@ import Rating from 'react-rating';
 import Avatar from './Avatar';
 
 import {isAuthenticated, isAdmin} from '../utils/auth';
-import {USER_TYPE_DEVELOPER} from '../constants/Api';
 
 export default class UserCardProfile extends React.Component {
   render() {
@@ -17,13 +16,9 @@ export default class UserCardProfile extends React.Component {
         </div>
         <div className="media-body">
           {profileLink
-            ? <div>
-                {user.type == USER_TYPE_DEVELOPER?(
-                  <a href={`/developer/${user.username}/`} target="_blank">{user.display_name}</a>
-                ):(
-                  <Link to={`/people/${user.username}/`} className="btn btn-block ">{user.display_name}</Link>
-                )}
-              </div>
+            ? <Link to={`/people/${user.username}/`}>
+                {user.display_name}
+              </Link>
             : <div>
                 {user.display_name}
               </div>}
